requirements.txt
azure-functions
Flask==1.1.1
